brand new test suite test

+++ test/stubs.h
@@ -1,4 +1,5 @@
-#include <stddef.h>
+#include <cstddef>
+#include <cstdlib>
 
 static void* crfStub_AlwaysNullMalloc(size_t size) {
     return NULL;
@@ -8,13 +9,13 @@ static void* crfStub_AlwaysNullRealloc(void* ptr, size_t size) {
 }
 
 static void* crfStub_OddNullMalloc(size_t size) {
-    static int state = 0;
+    static size_t state = 0;
     state = !state;
     if (state) return malloc(size);
     return NULL;
 }
 static void* crfStub_OddNullRealloc(void* ptr, size_t size) {
-    static int state = 0;
+    static size_t state = 0;
     state = !state;
     if (state) return realloc(ptr, size);
     return NULL;
